local/grpc_version/requirements.txt
grpcio
grpcio-tools
matplotlib
Pillow
numpy<2.0
opencv-python-headless
rembg
onnxruntime
torch
torchvision
torchaudio
requests
rich
openpyxl
